requirements.txt
app==0.0.1
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.4
Django==5.2
django-colorfield==0.14.0
django-crispy-forms==2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
packaging==25.0
pigame==1.9.5.dev0
pillow==11.2.1
pygame==2.6.1
pygame.snake==0.0.1a0
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
